add tests for controller and rename package
