+++ .graphqlconfig
@@ -1,0 +1,4 @@
+{
+  "name": "blue-steel schema",
+  "extensions": {}
+}
