+++ src/main/java/blue/steel/backend/story/Campaign.java
@@ -1,4 +1,4 @@
-package blue.steel.backend.campaing;
+package blue.steel.backend.story;
 
 import lombok.Data;
