+++ src/main/resources/graphql/story/actualCampaign.graphql
@@ -1,0 +1,8 @@
+query actualCampaign {
+    actualCampaign {
+        id
+        name
+        description
+        imageUrl
+    }
+}
