+++ src/main/resources/graphql/story/campaign.graphqls
