--- src/test/java/blue/steel/backend/campaign/CampaignControllerTest.java
@@ -1,2 +1,0 @@
-package blue.steel.backend.campaign;public class CampaignControllerTest {
-}
