+++ src/test/resources/application.properties
@@ -1,0 +1,1 @@
+logging.level.org.springframework.web=DEBUG
